Move custom snippets listing view tests to test_viewset and merge FilterableSnippet with FullFeaturedSnippet
